requirements.txt
openai-agents
openai
python-dotenv
aiofiles
requests
beautifulsoup4 